nbody docs
