+++ src/physics/fallingsand/elements/element.rs
@@ -10,8 +10,7 @@ use crate::physics::fallingsand::util::vectors::JkVector;
 use crate::physics::heat::components::{
     HeatEnergy, Length, SpecificHeat, ThermalConductivity, ThermodynamicTemperature,
 };
-use crate::physics::orbits::components::Mass;
-use crate::physics::orbits::nbody::Force;
+use crate::physics::orbits::components::{Force, Mass};
 use crate::physics::util::clock::Clock;
 use bevy::render::color::Color;
 use ndarray::Array2;
