+++ src/physics/orbits/components.rs
@@ -4,6 +4,8 @@
 use bevy::{ecs::component::Component, math::Vec2};
 use derive_more::{Add, AddAssign, Sub, SubAssign, Sum};
 
+use super::nbody::G;
+
 /// Indicates that an entity emits a gravitational field.
 #[derive(Component, Default, Debug, Clone, Copy)]
 pub struct GravitationalField;
@@ -15,3 +17,35 @@ pub struct Mass(pub f32);
 /// The velocity of an entity in meters per second.
 #[derive(Component, Debug, Clone, Copy, Add, Sub, AddAssign, SubAssign)]
 pub struct Velocity(pub Vec2);
+
+/// The force applied to an entity with its direction $\vec{N}$
+#[derive(Component, Debug, Clone, Copy)]
+pub struct ForceVec(pub Vec2);
+
+/// The scalar force applied to an entity in Newtons $N$
+#[derive(Component, Debug, Clone, Copy)]
+pub struct Force(pub f32);
+
+impl From<ForceVec> for Force {
+    fn from(force_vec: ForceVec) -> Self {
+        Force(force_vec.0.length())
+    }
+}
+
+impl Force {
+    /// Returns the force applied to the entity by gravitation
+    pub fn from_mass(mass: Mass, acceleration: GravitationalAcceleration) -> Self {
+        Force(mass.0 * acceleration.0)
+    }
+}
+
+/// The acceleration due to gravity
+#[derive(Component, Debug, Clone, Copy)]
+pub struct GravitationalAcceleration(pub f32);
+
+impl GravitationalAcceleration {
+    /// Returns the acceleration due to gravity towards a mass
+    pub fn from_total_mass(total_mass: Mass) -> Self {
+        GravitationalAcceleration(G * total_mass.0)
+    }
+}
