+++ src/physics/orbits/nbody.rs
@@ -1,7 +1,51 @@
+//! # NxM-Body Physics
+//!
+//! In this module, we implement a type of gravitational physics simulation I will call
+//! "NxM-Body" physics.
+//!
+//! Similar to N-Body physics, this models a more complex system of gravity than mere
+//! two-body physics. However, it is not as complex as N-Body physics, which models the
+//! gravitational interactions between all bodies in the system. Instead, NxM-Body physics
+//! models the gravitational interactions between all bodies in the system that have a
+//! gravitational field, which will likely just be the planets, moons, and stars, likely
+//! less than 100 bodies. We can call this N. The other bodies, which will likely be
+//! spacecraft, asteroids, and other small bodies, will be modeled as if they are affected
+//! by the gravitational fields of the bodies with gravitational fields, but not by each
+//! other. We can call this M.
+//!
+//! This is a simple model that is much less computationally expensive than N-Body physics,
+//! which is worst-case O(N^2), wheras NxM-Body physics is worst-case O(N*M). Since M is
+//! constant and small in any given simulation, this is effectively O(N).
+//!
+//! # Leapfrog Integration
+//!
+//! <https://en.wikipedia.org/wiki/Leapfrog_integration>
+//!
+//! We use the leapfrog integration method to update the positions and velocities of the
+//! bodies. This is a [symplectic integrator](https://en.wikipedia.org/wiki/Symplectic_integrator)
+//! , which means it conserves energy. This is
+//! important for a physics simulation, because it means that the simulation will not
+//! "drift" over time, and the bodies will not gain or lose energy over time.
+//!
+//! It is also a second-order integrator, which means that it is more accurate than a
+//! first-order integrator, like Euler's method. This means that the simulation will be
+//! more accurate, and will not require as small of a time step to be stable.
+//!
+//! It is acomplished by updating the velocity of the bodies by half a time step, then
+//! updating the position of the bodies by a full time step, then updating the velocity
+//! of the bodies by half a time step again.
+//!
+//! # Decision to use the CPU
+//!
+//! I originally wanted to use the GPU for this simulation, because it is a very parallel
+//! problem. However, I decided to use the CPU because it is a simpler solution, and each
+//! frame I would need to read the positions and velocities of all the bodies from the GPU
+//! which is actually slower than just doing the computation on the CPU. Using rayon, I can
+//! parallelize the computation on the CPU, so it is not a huge performance hit.
+
 use bevy::{
     app::{App, Plugin, Update},
     ecs::{
-        component::Component,
         entity::Entity,
         query::{With, Without},
         schedule::IntoSystemConfigs,
@@ -12,107 +56,74 @@ use bevy::{
     transform::components::Transform,
 };
 
-use super::components::{GravitationalField, Mass, Velocity};
+use super::components::{ForceVec, GravitationalField, Mass, Velocity};
 
 /// It's important that we don't compute the gravitational force between two bodies that are too
 /// close together, because the force will be very large and the simulation will be unstable.
+/// This is squared because then we don't have to do a square root to compare it to the distance squared in the force calculation.
 pub const MIN_DISTANCE_SQUARED: f32 = 100.0;
+
 /// The gravitational constant
-/// N⋅m^2/kg^2
+///
+/// $ \frac{N \cdot m^2}{kg^2} $
+///
+/// Note this has nothing to do with real life, and is just a scaling
+/// factor to make the simulation
+/// act at the scale of gravity we want.
 pub const G: f32 = 1.0e3;
 
-/// The force applied to an entity with its direction
-#[derive(Component, Debug, Clone, Copy)]
-pub struct ForceVec(pub Vec2);
+/// Returns the gravitational force between two entities
+fn compute_gravitational_force(
+    pos1: &Transform,
+    mass1: &Mass,
+    pos2: &Transform,
+    mass2: &Mass,
+) -> ForceVec {
+    let r = pos2.translation - pos1.translation;
+    let mut distance_squared = r.length_squared();
+    distance_squared = distance_squared.max(MIN_DISTANCE_SQUARED);
 
-/// The force applied to an entity
-#[derive(Component, Debug, Clone, Copy)]
-pub struct Force(pub f32);
+    // The gravitational constant G and masses are factored into the force magnitude
+    let force_magnitude = G * mass1.0 * mass2.0 / distance_squared;
 
-/// The force applied to an entity
-impl From<ForceVec> for Force {
-    fn from(force_vec: ForceVec) -> Self {
-        Force(force_vec.0.length())
-    }
+    // Calculate the force direction
+    // Normalize the displacement vector (r) to get the direction
+    let force_direction = r.normalize();
+
+    // The final force vector is the direction scaled by the force magnitude
+    ForceVec((force_direction * force_magnitude).xy())
 }
 
-impl Force {
-    /// Returns the force applied to the entity by gravitation
-    pub fn from_mass(mass: Mass, acceleration: GravitationalAcceleration) -> Self {
-        Force(mass.0 * acceleration.0)
-    }
-}
-
-/// The acceleration due to gravity
-#[derive(Component, Debug, Clone, Copy)]
-pub struct GravitationalAcceleration(pub f32);
-
-impl GravitationalAcceleration {
-    /// Returns the acceleration due to gravity towards a mass
-    pub fn from_total_mass(total_mass: Mass) -> Self {
-        GravitationalAcceleration(G * total_mass.0)
-    }
-}
-
-/// Just a namespace for the fundamental gravity functions
-struct GravityCalculations;
-
-impl GravityCalculations {
-    pub fn compute_gravitational_force(
-        pos1: &Transform,
-        mass1: &Mass,
-        pos2: &Transform,
-        mass2: &Mass,
-    ) -> ForceVec {
-        let r = pos2.translation - pos1.translation;
-        let mut distance_squared = r.length_squared();
-        distance_squared = distance_squared.max(MIN_DISTANCE_SQUARED);
-
-        // The gravitational constant G and masses are factored into the force magnitude
-        let force_magnitude = G * mass1.0 * mass2.0 / distance_squared;
-
-        // Calculate the force direction
-        // Normalize the displacement vector (r) to get the direction
-        let force_direction = r.normalize();
-
-        // The final force vector is the direction scaled by the force magnitude
-        ForceVec((force_direction * force_magnitude).xy())
-    }
-
-    /// Updates the velocity of the entity one half step
-    pub fn half_step_velocity_update(
-        this_body: (Entity, &Transform, &mut Velocity, &Mass),
-        other_bodies: &[(Entity, Transform, Velocity, Mass)],
-        dt: f32,
-    ) {
-        let mut net_force = Vec2::ZERO;
-        for other_body in other_bodies {
-            if this_body.0 == other_body.0 {
-                continue;
-            }
-            let force = Self::compute_gravitational_force(
-                this_body.1,
-                this_body.3,
-                &other_body.1,
-                &other_body.3,
-            );
-            net_force += force.0;
+/// Updates the velocity of the entity one half step
+fn half_step_velocity_update(
+    this_body: (Entity, &Transform, &mut Velocity, &Mass),
+    other_bodies: &[(Entity, Transform, Velocity, Mass)],
+    dt: f32,
+) {
+    let mut net_force = Vec2::ZERO;
+    for other_body in other_bodies {
+        if this_body.0 == other_body.0 {
+            continue;
         }
-        let vdiff = net_force / this_body.3 .0 * (dt / 2.0);
-        this_body.2 .0 += vdiff;
+        let force =
+            compute_gravitational_force(this_body.1, this_body.3, &other_body.1, &other_body.3);
+        net_force += force.0;
     }
+    let vdiff = net_force / this_body.3 .0 * (dt / 2.0);
+    this_body.2 .0 += vdiff;
+}
 
-    /// Updates the position of the entity one full step
-    pub fn full_position_update(this_body: (Entity, &mut Transform, &Velocity, &Mass), dt: f32) {
-        let pdiff = (this_body.2 .0 * dt).extend(0.0);
-        this_body.1.translation += pdiff;
-    }
+/// Updates the position of the entity one full step
+fn full_position_update(this_body: (Entity, &mut Transform, &Velocity, &Mass), dt: f32) {
+    let pdiff = (this_body.2 .0 * dt).extend(0.0);
+    this_body.1.translation += pdiff;
 }
 
 /// Plugin to set up nbody physics
 pub struct NBodyPlugin;
 
 impl Plugin for NBodyPlugin {
+    /// Adds the systems for the plugin
     fn build(&self, app: &mut App) {
         app.add_systems(
             Update,
@@ -126,6 +137,8 @@ impl Plugin for NBodyPlugin {
 
 /// Systems for the plugin
 impl NBodyPlugin {
+    /// Updates the locations and velocities of the entities with gravitational fields
+    /// based on each other entity with a gravitational field
     fn grav_bodies_system(
         mut grav_bodies: Query<
             (Entity, &mut Transform, &mut Velocity, &Mass),
@@ -141,16 +154,13 @@ impl NBodyPlugin {
         grav_bodies
             .par_iter_mut()
             .for_each(|(entity, mut transform, mut velocity, mass)| {
-                GravityCalculations::half_step_velocity_update(
+                half_step_velocity_update(
                     (entity, &transform, &mut velocity, mass),
                     &grav_bodies_copy,
                     dt,
                 );
-                GravityCalculations::full_position_update(
-                    (entity, &mut transform, &velocity, mass),
-                    dt,
-                );
-                GravityCalculations::half_step_velocity_update(
+                full_position_update((entity, &mut transform, &velocity, mass), dt);
+                half_step_velocity_update(
                     (entity, &transform, &mut velocity, mass),
                     &grav_bodies_copy,
                     dt,
@@ -158,6 +168,8 @@ impl NBodyPlugin {
             });
     }
 
+    /// Updates the locations and velocities of the entities without gravitational fields
+    /// based on the entities with gravitational fields
     fn no_grav_bodies_system(
         mut no_grav_bodies: Query<
             (Entity, &mut Transform, &mut Velocity, &Mass),
@@ -177,16 +189,13 @@ impl NBodyPlugin {
         no_grav_bodies
             .par_iter_mut()
             .for_each(|(entity, mut transform, mut velocity, mass)| {
-                GravityCalculations::half_step_velocity_update(
+                half_step_velocity_update(
                     (entity, &transform, &mut velocity, mass),
                     &grav_bodies_copy,
                     dt,
                 );
-                GravityCalculations::full_position_update(
-                    (entity, &mut transform, &velocity, mass),
-                    dt,
-                );
-                GravityCalculations::half_step_velocity_update(
+                full_position_update((entity, &mut transform, &velocity, mass), dt);
+                half_step_velocity_update(
                     (entity, &transform, &mut velocity, mass),
                     &grav_bodies_copy,
                     dt,
